% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_bands.R
\name{conf_bands}
\alias{conf_bands}
\title{Confidence bands for forecast comparisons.}
\usage{
conf_bands(
  scores,
  metric = "SS",
  m_bench = length(scores),
  type = "bonferroni",
  B = 1000,
  l = 3 * floor(length(scores[[1]])^(1/4)),
  alpha = 0.05
)
}
\arguments{
\item{scores}{A nested list containing the scores.The structure has to be the same for all forecasting methods:
\itemize{
\item{First dimension: Forecasting method m=1,...,M. Even if M=1, use a list structure.}
\item{Second dimension: Time points of the evaluation sample t=1,...,N.}
\item{Additional dimensions: The amount and the order of additional dimensions can be chosen arbitrarily, for example variables, locations, forecast horizons.}
}}

\item{metric}{A character object that determines which metric should be estimated from the scores. Allowed are \code{"ES"} (expected scores), \code{"SS"} (skill scores), \code{"RA"} (relative accuracy), and \code{"D"} (difference in expected scores). Default is \code{"SS"}. Except for \code{"ES"}, the metric is estimated given the benchmark method \code{m_bench}.}

\item{m_bench}{Integer between \code{1} and \code{M} that indicates the benchmark method. Default is \code{M}. Parameter has no effect if \code{metric = "ES"}.}

\item{type}{Type of confidence band: \code{c("bonferroni", "sup-t", "pointwise")}. Default is \code{"bonferroni"}.}

\item{B}{Number of bootstrap replications. Default is 1000.}

\item{l}{Block length used in the moving block bootstrap. Default is \code{3*floor(N^(1/4))}, where \code{N} is the size of the evaluation sample.}

\item{alpha}{Significance level. Default is \code{0.05}.}
}
\value{
A nested list containing a vector of length 3 (lower confidence bound, the estimated metric, and the upper bound). The list has the same structure as the input except for:
\itemize{
\item{The first dimension only includes the remaining forecasting methods, i.e. it is of length \code{M} if \code{metric = "ES"}, and of length \code{M-1} otherwise.}
\item{There is no time dimension. Thus, the second dimension of the output list is equal to the third dimension of the input list.}
\item{A vector is given at the last level instead of a single number.}
}
}
\description{
Estimate the skill scores, expected scores, relative accuracy or score differences and generate simultaneous confidence bands of level \code{1-alpha}.
}
