% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert list to data frame.}
\usage{
list_to_df(list, dimnames, dims_numeric = NULL, last_level = "vector")
}
\arguments{
\item{list}{A \emph{named} list that may be nested.}

\item{dimnames}{A character string or a vector of character strings stating the desired column names. The order has to align with the structure of the nested list.}

\item{dims_numeric}{Optional. A character string or a vector of character strings that contains the columns specified in \code{dimnames} that should be turned into a numeric column.}

\item{last_level}{If equal to \code{"vector"} (default), the very last level of the nested list is a vector. The default is suitable when applying the function to the output of \code{conf_bands}. Specifying \code{last_level = "value"} is appropriate when the nested list contains a numeric value at the last level.}
}
\value{
A data frame containing all columns specified
}
\description{
Convert a named list that is possibly nested to a data frame. Can be applied to the output of \code{conf_bands}, given that the input list had names.
}
