% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_horizontal.R
\name{plot_horizontal}
\alias{plot_horizontal}
\title{Plot horizontal confidence bands.}
\usage{
plot_horizontal(
  df,
  xvar,
  y = "SS",
  y_LB = "LB_bs",
  y_UB = "UB_bs",
  ylim = NULL,
  xlim = NULL,
  color_by = NULL,
  facet_by = NULL,
  facet_cols = NULL,
  markzero = TRUE,
  ylab = NULL,
  xlab = NULL,
  xbreaks = NULL,
  plottitle = NULL,
  theme = NULL
)
}
\arguments{
\item{df}{a data frame containing the point estimate and the lower and upper bounds.}

\item{xvar}{Name of the variable on the x-axis.}

\item{y}{Name of the point estimate. Default is \code{SS}.}

\item{y_LB}{Name of the lower bound. Default is \code{LB_bs} as produced by the \code{conf_bands}function of our package.}

\item{y_UB}{Name of the upper bound. Default is \code{UB_bs} as produced by the \code{conf_bands}function of our package.}

\item{ylim}{Optional vector of length 2, containing the limits for the y-axis of the plot.}

\item{xlim}{Optional vector of length 2, containing the limits for the x-axis of the plot.}

\item{color_by}{Optional character string. Use different colors depending on the variable name supplied to this parameter. For example, choose different colors depending on the method.}

\item{facet_by}{Optional character string. Use different facets depending on the variable name supplied to this parameter. For example, create facts for different variables of interest.}

\item{facet_cols}{Optional. The number of columns used for the facets.}

\item{markzero}{Optional. If \code{TRUE} (default), plot black horizontal line at \code{y = 0}. If \code{FALSE}, no line will be added.}

\item{ylab}{Optional character string containing the label of the y-axis.}

\item{xlab}{Optional character string containing the label of the x-axis.}

\item{xbreaks}{Optional. Number of breaks on the x-axis.}

\item{plottitle}{Optional character string containing the plot title.}

\item{theme}{Optional theme of the plot, e.g., theme_bw() or some custom theme.}
}
\value{
A plot containing the point estimates with horizontal confidence bands.
}
\description{
Generate a plot of the point estimates with horizontal confidence bands, e.g., over different forecast horizons, potentially for different variables.
}
