% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vertical.R
\name{plot_vertical}
\alias{plot_vertical}
\title{Plot vertical confidence bands.}
\usage{
plot_vertical(
  df,
  xvar = NULL,
  y = "SS",
  y_LB = "LB_bs",
  y_UB = "UB_bs",
  color_by = NULL,
  markzero = TRUE,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  plottitle = NULL,
  theme = NULL
)
}
\arguments{
\item{df}{a data frame containing the point estimate and the lower and upper bounds.}

\item{xvar}{Name of the variable on the x-axis (do not use quotation marks).}

\item{y}{Name of the point estimate (do not use quotation marks). Default is \code{SS}.}

\item{y_LB}{Name of the lower bound (do not use quotation marks). Default is \code{LB_bs} as produced by the \code{conf_bands}function of our package.}

\item{y_UB}{Name of the upper bound (do not use quotation marks). Default is \code{UB_bs} as produced by the \code{conf_bands}function of our package.}

\item{color_by}{Optional. Use different colors depending on the variable name (without quotation marks) supplied to this parameter, for example, choose different colors depending on the method.}

\item{markzero}{Optional. If \code{TRUE} (default), plot black horizontal line at \code{y = 0}. If \code{FALSE}, no line will be added.}

\item{ylim}{Optional vector of length 2, containing the limits for the y-axis of the plot.}

\item{xlab}{Optional character string containing the label of the x-axis.}

\item{ylab}{Optional character string containing the label of the y-axis.}

\item{plottitle}{Optional character string containing the plot title.}

\item{theme}{Optional theme of the plot, e.g., theme_bw() or some custom theme.}
}
\value{
A plot containing the point estimates with vertical confidence bands.
}
\description{
Generate a plot of the point estimates with vertical confidence bands, e.g. for different variables.
}
