% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UQforecasts-package.R
\docType{package}
\name{UQforecasts-package}
\alias{UQforecasts}
\alias{UQforecasts-package}
\title{UQforecasts: Uncertainty Quantification in Forecast Comparisons}
\description{
The package generates simultaneous confidence bands for skill scores, relative expected scores, expected score differences, and expected scores themselves. They allow for valid uncertainty quantification and provide results to joint equal predictive accuracy tests while preventing multiple comparison problems.
}
\author{
\strong{Maintainer}: Tanja Zahn \email{tzahn@wiwi.uni-frankfurt.de}

Authors:
\itemize{
  \item Marc-Oliver Pohle
  \item Sebastian Lerch
}

}
\keyword{internal}
